+++ CHANGELOG.md
@@ -2,8 +2,19 @@
 
 All notable changes to the "jdl-noml-uml" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [1.1.0] - 2025-02-12
 
-## [Unreleased]
+### Added
+
+- Language icons.
+- Extension icon.
+- Language configuration for JDL.
+
+### Changed
+
+- Improve JDL syntaxe highlighting.
+- Disallow to run Nomnoml viewer on files that are neither JDL nor Nomnoml files.
+
+## [1.0.0] - 2025-02-09
 
 - Initial release